pyproject.toml
[project]
name = "muz-bot"
version = "0.2.0"
description = "muz-bot for QQ"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "nonebot2[fastapi]",
    "nonebot-adapter-onebot",
    "Pillow",
    "httpx",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["buaa-by", "menu", "cf_plugin", "signin", "duaa_plugin", "pjsk_plugin"]  # 以后你新建了插件，名字记得加在这里！
plugin_dirs = []
builtin_plugins = ["echo", "nonebot_plugin_apscheduler"]
